Call health check before sending data to Open Podcast API
Since the API runs on Google cloud run at the moment
and can scale down to 0 instances if not needed,
there can be a delay in the availability of the service
in a cold-start scenario. To avoid errors, we wait until
the health check responds with a `200` status code
before we continue.