Harden code and make it more Pythonic (#36)
I've used `pylint` to fix some issues with the code.
This should not change the pipeline logic.

To make sure I ran the changes locally (running the dev api
and dev db this time) and the data in Planetscale looks okay
and I got no API errors. Also checked the `updates` table.